added bodies for the windows dirent functions

+++ viper/include/viper/core/debug/exceptions.h
@@ -16,12 +16,12 @@
 #define ViperThrow(NUM) \
    __ViperJump(NUM)
 
-extern i32 __ViperJumpIndexDecrease(i32 i);
-extern i32 __ViperJumpIndexIncrease(i32 i);
+extern i64 __ViperJumpIndexDecrease(i64 i);
+extern i64 __ViperJumpIndexIncrease(i64 i);
 extern sigjmp_buf* __ViperJumpGetBuffer();
 extern sigjmp_buf* __ViperJumpGetNextBuffer();
 extern void __ViperJump(i32 error);
 
-extern i32 __ViperJumpSignalHandlerInit();
+extern i8 __ViperJumpSignalHandlerInit();
 
 #endif /* __header_viper_core_debug_exceptions__ */
