+++ viper/source/core/debug/exceptions.c
@@ -17,7 +17,7 @@ static void __ViperJumpSignalHandler(int signal, siginfo_t* action, void* data) 
    }
 }
 
-inline i64 __ViperJumpIndexDecrease(i32 i) {
+inline i64 __ViperJumpIndexDecrease(i64 i) {
    return __viperJumpIndex -= i;
 }
 
@@ -27,7 +27,7 @@ inline i64 __ViperJumpIndexDecrease(i32 i) {
  * @param i : A integer value will be added
  * to the index.
  */
-inline i64 __ViperJumpIndexIncrease(i32 i) {
+inline i64 __ViperJumpIndexIncrease(i64 i) {
    return __viperJumpIndex += i;
 }
